LPD-90495 renames JiraAssetsClient to JiraAssetsService and moves to service package
